change `StrategyWeight` to be u32
this means that the entire type can fit in 64 bits making things go
vroom

+++ radguy/src/ordered/strategy/mod.rs
@@ -19,7 +19,7 @@ pub use orx::OrxStrategy;
 #[derive(PartialEq, Eq, Clone, Copy, Hash, Debug)]
 pub enum StrategyWeight {
     Infinity,
-    Num(u64),
+    Num(u32),
 }
 
 impl Ord for StrategyWeight {
